# swagger.yaml
openapi: 3.0.3
info:
  title: Wine Quality Prediction API
  description: |
    🍷 **API de prédiction de la qualité du vin**
    
    Cette API utilise un modèle de Machine Learning pour prédire la qualité d'un vin 
    en fonction de ses caractéristiques physico-chimiques.
    
    ## Fonctionnalités
    
    * **Prédiction** : Obtenir un score de qualité (0-10) pour un vin
    * **Auto-update** : Le modèle se met à jour automatiquement depuis MLflow
    * **Monitoring** : Health checks et informations sur le modèle
    
    ## Modèle
    
    Le modèle est entraîné sur le dataset Wine Quality et prédit un score entre 0 et 10.
  version: 1.0.0
  contact:
    name: Wine Quality Team
    email: support@winequality.com
  license:
    name: MIT

servers:
  - url: http://localhost:8000
    description: Serveur local de développement
  - url: http://fastapi:8000
    description: Serveur Docker

tags:
  - name: General
    description: Endpoints généraux
  - name: Monitoring
    description: Surveillance et health checks
  - name: Prediction
    description: Prédiction de la qualité du vin
  - name: Model Management
    description: Gestion du modèle ML

paths:
  /:
    get:
      tags:
        - General
      summary: Racine de l'API
      description: Retourne les informations de base sur l'API
      operationId: root
      responses:
        '200':
          description: Informations sur l'API
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Wine Quality Prediction API
                  status:
                    type: string
                    example: running
                  model_name:
                    type: string
                    example: wine-quality-model
                  docs:
                    type: string
                    example: /docs
                  health:
                    type: string
                    example: /health

  /health:
    get:
      tags:
        - Monitoring
      summary: Health check
      description: Vérifier le statut de l'API et la disponibilité du modèle
      operationId: health_check
      responses:
        '200':
          description: Statut de santé de l'API
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: healthy
                  model_loaded:
                    type: boolean
                    example: true
                  model_name:
                    type: string
                    example: wine-quality-model
                  model_version:
                    type: string
                    example: "3"

  /model/info:
    get:
      tags:
        - Model Management
      summary: Informations sur le modèle
      description: Obtenir les métadonnées détaillées du modèle actuellement chargé
      operationId: model_info
      responses:
        '200':
          description: Métadonnées du modèle
          content:
            application/json:
              schema:
                type: object
                properties:
                  model_name:
                    type: string
                    example: wine-quality-model
                  version:
                    type: string
                    example: "3"
                  run_id:
                    type: string
                    example: a1b2c3d4e5f6g7h8i9j0
                  status:
                    type: string
                    example: READY
                  creation_timestamp:
                    type: integer
                    example: 1699564800000
        '503':
          description: Modèle non chargé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
              example:
                detail: Model not loaded. Please wait for model initialization or train a new model.

  /predict:
    post:
      tags:
        - Prediction
      summary: Prédire la qualité du vin
      description: |
        Prédire la qualité d'un vin (score 0-10) à partir de ses caractéristiques physico-chimiques.
        
        Le modèle vérifie automatiquement s'il existe une nouvelle version avant chaque prédiction.
      operationId: predict
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WineFeatures'
            examples:
              white_wine:
                summary: Vin blanc
                value:
                  type: white
                  fixed_acidity: 7.0
                  volatile_acidity: 0.27
                  citric_acid: 0.36
                  residual_sugar: 20.7
                  chlorides: 0.045
                  free_sulfur_dioxide: 45.0
                  total_sulfur_dioxide: 170.0
                  density: 1.001
                  pH: 3.0
                  sulphates: 0.45
                  alcohol: 8.8
              red_wine:
                summary: Vin rouge
                value:
                  type: red
                  fixed_acidity: 7.4
                  volatile_acidity: 0.70
                  citric_acid: 0.00
                  residual_sugar: 1.9
                  chlorides: 0.076
                  free_sulfur_dioxide: 11.0
                  total_sulfur_dioxide: 34.0
                  density: 0.9978
                  pH: 3.51
                  sulphates: 0.56
                  alcohol: 9.4
      responses:
        '200':
          description: Prédiction réussie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PredictionResponse'
        '503':
          description: Modèle non disponible
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
              example:
                detail: Model 'wine-quality-model' not loaded. Please train and register a model first.
        '500':
          description: Erreur interne
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'

  /model/reload:
    post:
      tags:
        - Model Management
      summary: Recharger le modèle
      description: Forcer le rechargement manuel de la dernière version du modèle depuis MLflow
      operationId: reload_model
      responses:
        '200':
          description: Modèle rechargé avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Model reloaded successfully
                  model_version:
                    type: string
                    example: "3"
                  model_name:
                    type: string
                    example: wine-quality-model
        '404':
          description: Modèle non trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
              example:
                detail: No model 'wine-quality-model' found in MLflow
        '500':
          description: Erreur lors du rechargement
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
  
  /matthias:
    get:
      tags:
        - General
      summary: Petit clin d'oeil/easter egg de Matthias (Second degré, bien sûr !)
      description: Retourne une vidéo de Matthias
      operationId: matthias_endpoint
      responses:
        '200':
          description: 
          content:
            video/mp4:
              schema:
                type: string
                format: binary
        '500':
          description: Erreur interne
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'

components:
  schemas:
    WineFeatures:
      type: object
      required:
        - type
        - fixed_acidity
        - volatile_acidity
        - citric_acid
        - residual_sugar
        - chlorides
        - free_sulfur_dioxide
        - total_sulfur_dioxide
        - density
        - pH
        - sulphates
        - alcohol
      properties:
        type:
          type: string
          enum:
            - red
            - white
          description: Type de vin
          example: white
        fixed_acidity:
          type: number
          format: float
          minimum: 0
          description: Acidité fixe (g/L)
          example: 7.0
        volatile_acidity:
          type: number
          format: float
          minimum: 0
          description: Acidité volatile (g/L)
          example: 0.27
        citric_acid:
          type: number
          format: float
          minimum: 0
          description: Acide citrique (g/L)
          example: 0.36
        residual_sugar:
          type: number
          format: float
          minimum: 0
          description: Sucre résiduel (g/L)
          example: 20.7
        chlorides:
          type: number
          format: float
          minimum: 0
          description: Chlorures (g/L)
          example: 0.045
        free_sulfur_dioxide:
          type: number
          format: float
          minimum: 0
          description: SO2 libre (mg/L)
          example: 45.0
        total_sulfur_dioxide:
          type: number
          format: float
          minimum: 0
          description: SO2 total (mg/L)
          example: 170.0
        density:
          type: number
          format: float
          minimum: 0
          description: Densité (g/cm³)
          example: 1.001
        pH:
          type: number
          format: float
          minimum: 0
          maximum: 14
          description: pH
          example: 3.0
        sulphates:
          type: number
          format: float
          minimum: 0
          description: Sulfates (g/L)
          example: 0.45
        alcohol:
          type: number
          format: float
          minimum: 0
          maximum: 20
          description: Alcool (% vol)
          example: 8.8

    PredictionResponse:
      type: object
      properties:
        quality_prediction:
          type: number
          format: float
          description: Score de qualité continu (0-10)
          example: 6.234
        quality_class:
          type: integer
          description: Score de qualité arrondi (0-10)
          example: 6
        model_updated:
          type: boolean
          description: Indique si le modèle a été mis à jour lors de cette prédiction
          example: false
        model_version:
          type: string
          description: Version du modèle utilisé
          example: "3"
        model_name:
          type: string
          description: Nom du modèle
          example: wine-quality-model
        wine_type:
          type: string
          description: Type de vin (répété pour confirmation)
          example: white

    HTTPError:
      type: object
      properties:
        detail:
          type: string
          description: Message d'erreur détaillé
          example: Model not loaded